feat: DELETE_ALL

+++ table.py
@@ -1,4 +1,5 @@
 import glob
+import json
 import os
 import time
 
@@ -71,6 +72,11 @@ def removeColumnFamily(table_name, column_family):
 
 def drop_table(table_name):
     try:
+        with open(f"db/{table_name}.hfile.json", 'r') as file:
+            table = json.load(file)
+        if table['enabled']:
+            print(f"Error: La tabla '{table_name}' está habilitada.")
+            return
         os.remove(f"db/{table_name}.hfile.json")
     except FileNotFoundError:
         print(f"Error: La tabla '{table_name}' no existe.")
@@ -80,6 +86,11 @@ def drop_table(table_name):
 def drop_all_tables():
     tables = glob.glob('db/*.hfile.json')
     for table in tables:
+        with open(table, 'r') as file:
+            table_data = json.load(file)
+        if table_data['enabled']:
+            print(f"Error: La tabla '{os.path.basename(table)}' está habilitada.")
+            continue
         os.remove(table)
     print(f"{len(tables)} tablas eliminadas.")
 
@@ -93,3 +104,5 @@ def get(table_name, row_key):
     table.get(row_key)
     end = time.time()
     print(f"Took {end - start:.4f} seconds")
+
+
